requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
faiss-cpu
pypdf
python-dotenv
openai
